+++ client/views/School/Test/View.jsx
@@ -7,40 +7,13 @@ import SchoolTestHeader from './Header.jsx';
 import SchoolTestActions from './Actions.jsx';
 import SchoolTestForm from './Form/index.jsx';
 
-SchoolTestView = React.createClass({
-  mixins: [ReactMeteorData],
-
-  /* Reactive Data Fetching
-  */
-
-  getMeteorData() {
-    const { testId } = this.props;
-
-    const handles = {
-      test: Meteor.subscribe('SchoolTests', { testId },
-        { subjects: true, tags: true, questions: true, course: true }),
-      subjects: Meteor.subscribe('PublicSubjects'),
-    };
-
-    const data = {
-      ready: _.mapValues(handles, h => h.ready()),
-      subjects: Fetch.Public().subjects().fetch(),
-      test: _.first(Fetch.General.tests(testId).fetch()),
-    };
-
-    data.tags = data.test && data.test.findTags().fetch();
-    data.questions = data.test && data.test.findQuestions().fetch();
-    data.course = data.test && data.test.findCourse().fetch();
-    data.documents = data.test && _.first(data.test.findDocuments().fetch());
-
-    return data;
-  },
+const SchoolTestView = React.createClass({
 
   /* Render
   */
 
   render() {
-    const { ready, test, subjects } = this.data;
+    const { ready, test, subjects } = this.props;
     return (
       <div className='ui container'>
 
@@ -50,12 +23,12 @@ SchoolTestView = React.createClass({
 
         <Paper>
           {!_.every(ready) ? <LinearProgress/> : [
-            <SchoolTestHeader key='header' {...this.data} />,
+            <SchoolTestHeader key='header' {...this.props} />,
             <Divider key='d0'/>,
             <SchoolTestForm key='question' doc={test} subjects={subjects} />,
 
             // <Divider key='d1'/>,
-            // <SchoolTestActions key='actions' {...this.data} />,
+            // <SchoolTestActions key='actions' {...this.props} />,
           ]}
         </Paper>
 
@@ -63,3 +36,5 @@ SchoolTestView = React.createClass({
     );
   },
 });
+
+export default SchoolTestView;
